remove duplicated hdwallet file and add some basic tests from it
